scripts: tooling: split vuart code into support files
Split vuart code into a support file separate from the tt-console logic,
so other tools can interact with the vuart.
